lazer/pyproject.toml
[tool.poetry]
name = "lazer"
version = "0.1.0"
description = ""
authors = ["Justin Stitt <jstitt007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
icecream = "^2.1.3"
pytojsonschema = "^1.11.1"
pydantic = "^1.10.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
